Moving test apps to src/tests
