+++ app/bergamot.cpp
@@ -1,5 +1,4 @@
 #include "cli.h"
-#include "testapps.h"
 
 int main(int argc, char *argv[]) {
   auto cp = marian::bergamot::createConfigParser();
@@ -12,24 +11,6 @@ int main(int argc, char *argv[]) {
     app::native(options);
   } else if (mode == "decoder") {
     app::decoder(options);
-  }
-
-  else if (mode == "test-quality-scores") {
-    testapp::quality_scores(options);
-  } else if (mode == "test-alignment-scores") {
-    testapp::alignment_aggregated_to_source(options, /*numeric=*/true);
-  } else if (mode == "test-alignment-words") {
-    testapp::alignment_aggregated_to_source(options, /*numeric=*/false);
-  } else if (mode == "test-response-source-sentences") {
-    testapp::annotated_text_sentences(options, /*source=*/true);
-  } else if (mode == "test-response-target-sentences") {
-    testapp::annotated_text_sentences(options, /*source=*/false);
-  } else if (mode == "test-response-source-words") {
-    testapp::annotated_text_words(options, /*source=*/true);
-  } else if (mode == "test-response-target-words") {
-    testapp::annotated_text_words(options, /*source=*/false);
-  } else if (mode == "legacy-service-cli") {
-    testapp::legacy_service_cli(options);
   } else {
     ABORT("Unknown --mode {}. Use one of: {wasm,native,decoder}", mode);
   }
