+++ src/CMakeLists.txt
@@ -1,7 +1,6 @@
 add_subdirectory(translator)
 
-if(COMPILE_TESTS)
-  # Catch currently comes from marian sources.
-  add_subdirectory(tests)
+if (COMPILE_TESTS)
+    add_subdirectory(tests)
 endif(COMPILE_TESTS)
